requirements.txt
pygame-texteditor